renamed services and extracted interface
